Capstone 2 sandwich shop init push
in this initial push i set up a UML diagram and project structure. also added notes to each class to help me as i start to work on the project. notes include details for each class. Diagram and notes were done with the help of AI and multiple prompts to get some i liked. Structure and basic idea were done on my own